changed to single precision

+++ png2array/png2array.h
@@ -8,6 +8,6 @@
 #include "lodepng.h"
 
 void processData(const char *dataPath, int samplesPerLabel, int size, int labels, const char *outPath);
-void loadData(const char* path, double** &data, double* &mean, double* &deviation, int &labels, int &samplesPerLabels, int &featuresDimension);
+void loadData(const char* path, float** &data, float* &mean, float* &deviation, int &labels, int &samplesPerLabels, int &featuresDimension);
 int *genTargets(int labels, int samplesPerLabels);
-double *decodeTwoSteps(std::string filename, int &w, int &h);
+float *decodeTwoSteps(std::string filename, int &w, int &h);
